requirements/requirements_dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --no-hashes --python 3.11
-e .
-e ./packages/_dev
accessible-pygments==0.0.5
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.5.0
appnope==0.1.4 ; platform_system == 'Darwin'
astroid==3.3.2
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1 ; implementation_name == 'pypy'
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
coverage==7.6.1
cyclopts==2.9.3
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
docstring-parser==0.16
docutils==0.21.2
dulwich==0.22.1
executing==2.1.0
fastjsonschema==2.20.0
fawltydeps==0.16.0
filelock==3.16.1
greenlet==3.1.0 ; (python_full_version < '3.13' and platform_machine == 'AMD64') or (python_full_version < '3.13' and platform_machine == 'WIN32') or (python_full_version < '3.13' and platform_machine == 'aarch64') or (python_full_version < '3.13' and platform_machine == 'amd64') or (python_full_version < '3.13' and platform_machine == 'ppc64le') or (python_full_version < '3.13' and platform_machine == 'win32') or (python_full_version < '3.13' and platform_machine == 'x86_64')
h11==0.14.0
identify==2.6.1
idna==3.10
imagesize==1.4.1
importlib-metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-cache==1.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
latexcodec==3.0.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-nb==1.1.1
myst-parser==3.0.1
nbclient==0.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
numpydoc==1.7.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pip==24.2
pip-requirements-parser==32.0.1
pipdeptree==2.23.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.7.1
prompt-toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pure-eval==0.2.3
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22 ; implementation_name == 'pypy'
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pydata-sphinx-theme==0.15.4
pygments==2.18.0
pyparsing==3.1.4
pyright==1.1.371
pytest==8.2.2
pytest-cov==5.0.0
pytest-custom-exit-code==0.3.0
pytest-github-actions-annotate-failures==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rich-rst==1.3.1
rpds-py==0.20.0
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8 ; python_full_version < '3.13' and platform_python_implementation == 'CPython'
ruff==0.5.1
setuptools==75.1.0
six==1.16.0
snakeviz==2.2.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
sourcery==1.21.0
sphinx==7.3.7
sphinx-autobuild==2024.4.16
sphinx-autodoc2==0.5.0
sphinx-book-theme==1.1.3
sphinx-design==0.6.0
sphinx-thebe==0.3.1
sphinx-tippy==0.4.3
sphinx-togglebutton==0.3.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-towncrier==0.4.0a0
sqlalchemy==2.0.35
stack-data==0.6.3
starlette==0.38.5
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1 ; python_full_version <= '3.11'
tornado==6.4.1
towncrier==23.11.0
traitlets==5.14.3
typing-extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.30.6
virtualenv==20.26.5
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
wheel==0.44.0
zipp==3.20.2
